Use variable `value_cstr` to determine whether to return boolean `true` or `false` from function `ParseFlag` on completion
This commit fixes bug that made function `ParseFlag` return `true` in every case even if
the `flag_name` is not present in the user given flag.  By using the statement:

  return (*value = (std::strcmp(value_cstr, "true") == 0));

we ensure that function `ParseFlag` only returns `true` or `flase` if user had used that
argument or not used that argument respectively.

Immediately return from function `ParseXTestFlag` if parsing has taking place even for
one flag.  This will ensure that we don't overdo work.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>