FIX #5 -- Introduce new macro `XTEST_TESTING_DISABLED` to disable tests when building `xtest`'s tests together with the library
This macro is needed as out tests uses output redirection so to capture the output in a
custom buffer to later check it with the expected output.  As we know `ColoredPrintf()`
is the only gateway to our data to the console which internally uses function
`ShouldUseColor()` that is bound to return `false` if any redirection happens.

This commit also fixes the problem with flag `--color` which was not working before
because of the old macro `XTEST_TESTING_ENABLED` which we always forgot to define before
building the library with samples.

There are two ways to define macro `XTEST_TESTING_DISABLED`:
  * Explicitly provide `XTEST_TESTING_DISABLED` option over the command line while
    building the library.
  * Removing `BUILD_TESTS` option from the command line while building the library.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>