Add `XTEST_FLAG_PREFIX_` i.e., `xtest_` before every xtest flag over the command line
Modify function `ParseFlagValue()` to also take in account the `XTEST_FLAG_PREFIX_`
before every xtest flag over the command line.

This commit brings a important change where now it is easy to differentiate xtest's
flags with other flags in case xtest is used to test a project that also depends on
command line arguments, this way we only handle those commands that has a valid
`XTEST_FLAG_PREFIX_` at the front of them otherwise we leave such flags on the `argv`
array for later use by any other program.

This commit also fixes the spelling mistake that we made in `prefix`.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>