Introduce new command line argument `--xtest_list_tests` to list test names with their test suite names
Flag `--xtest_list_tests` lists the test names with their suite names with running them
which is a good feature when a user wants a quick lookup of the tests registered.

Function `ListTestsWithSuiteName()` helps in this quick lookup of registered tests using
library xtest with the following output format:

  ```
  InfixTest.
   TestWithElementsWithoutBraces
   TestWithElementsInsideBraces
   TestWithElementsEnclosedInBraces
  ```

This commit also fixes issue with template function `ParseFlag()` where the conditional
where we must abort the program if parsing fails should be

  if (*value_cstr == '\0')

not

  if (value_cstr == nullptr)

which is clearly a readable bug because function `ParseFlagValue()` never returns a
`nullptr` instead to denote a failure it returns a empty string.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>